refactored into worker class for modular use in other tools
